DAB radio daemon (dabd)
dabd is a DAB radio daemon which accepts commands from stdin
and sends its result strings to stdout. So you can control dabd
using the keyboard.
Using the named pipe (FIFO) mechanism of Linux offers a lot of
possibilities to redirect the DAB radio control to more convenient
user frontends as there are
* Graphical User Interface
* GPIO control
* Web server control via browser
* ...